refactor file names and highiligh
