--- lua/README.md
@@ -1,1 +1,0 @@
-The file `actions` has cross cutting functionalities
