+++ lua/vuffers.lua
@@ -1,7 +1,7 @@
-local ui = require("vuffers.ui")
+local window = require("vuffers.window")
 local bufs = require("vuffers.buffers")
 local auto_commands = require("vuffers.auto-commands")
-local actions = require("vuffers.actions")
+local ui = require("vuffers.ui")
 local key_bindings = require("vuffers.key-bindings")
 
 local M = {}
@@ -11,7 +11,7 @@ function M.setup(opts)
 end
 
 function M.toggle()
-  if ui.is_hidden() then
+  if window.is_hidden() then
     M.open()
   else
     M.close()
@@ -19,31 +19,32 @@ function M.toggle()
 end
 
 function M.open()
-  if not ui.is_hidden() then
+  if not window.is_hidden() then
     return
   end
 
-  local is_valid = ui.is_valid()
+  local is_valid = window.is_valid()
 
   auto_commands.create_auto_group()
 
   bufs.reload_all_buffers()
 
-  ui.open()
-  actions.render_buffers()
+  window.open()
+  ui.render_buffers()
+  ui.highlight_active_buffer()
 
   if not is_valid then
-    key_bindings.init(ui.get_split_buf_num())
+    key_bindings.init(window.get_split_buf_num())
   end
 end
 
 function M.close()
-  if ui.is_hidden() then
+  if window.is_hidden() then
     return
   end
 
   auto_commands.remove_auto_group()
-  ui.close()
+  window.close()
 end
 
 return M
