+++ lua/vuffers/auto-commands/buffers.lua
@@ -1,5 +1,5 @@
 local buffers = require("vuffers.buffers")
-local actions = require("vuffers.actions")
+local ui = require("vuffers.ui")
 
 local M = {}
 
@@ -7,20 +7,22 @@ local M = {}
 ---@param filetype string
 function M.on_buf_enter(buffer, filetype)
   buffers.set_current_bufnr(buffer, filetype)
-  actions.render_buffers()
+  ui.highlight_active_buffer()
 end
 
 ---@param buffer {buf: number, event: string, file: string, group: number, id: number, match: string}
 ---@param filetype string
 function M.on_buf_add(buffer, filetype)
   buffers.add_buffer(buffer, filetype)
-  actions.render_buffers()
+  ui.render_buffers()
+  ui.highlight_active_buffer()
 end
 
 ---@param buffer {buf: number, event: string, file: string, group: number, id: number, match: string}
 function M.on_buf_delete(buffer)
   buffers.remove_buffer(buffer.buf)
-  actions.render_buffers()
+  ui.render_buffers()
+  ui.highlight_active_buffer()
 end
 
 return M
